+++ app/src/main/java/com/infix/musicappv1/data/source/local/artist/ArtistDao.kt
@@ -32,7 +32,7 @@ interface ArtistDao {
         FROM artists
         WHERE artist_id = :artistId
     """)
-    fun getArtistById(artistId: Int): Artist?
+    fun getArtistById(artistId: Long): Artist?
 
     @Query(
         """
