+++ app/src/main/java/com/infix/musicappv1/ui/discovery/artist/detail/ArtistDetailFragment.kt
@@ -14,8 +14,8 @@ import com.infix.musicappv1.data.model.playlist.Playlist
 import com.infix.musicappv1.data.model.song.Song
 import com.infix.musicappv1.data.repository.PermissionRepository
 import com.infix.musicappv1.databinding.FragmentArtistDetailBinding
-import com.infix.musicappv1.ui.base.BasePlayMusicFragment
 import com.infix.musicappv1.ui.adapter.song.SongAdapter
+import com.infix.musicappv1.ui.base.BasePlayMusicFragment
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
@@ -86,7 +86,7 @@ class ArtistDetailFragment : BasePlayMusicFragment() {
                     playSong(
                         pos,
                         Playlist(
-                            playlistId = artist.id,
+                            playlistId = artist.id.toLong(),
                             namePlaylist = artist.name,
                             artwork = artist.avatar
                         ),
